[thenakliman] #1 Removed absolute url for backend service
Frontend service sends request relative to its own domain but
without apache, it is not possible to run both the service on
the same port. A temporary fix was introduced to workaround. This
patch removed the workaround.

Coming patches will be adding apache conf file for proxy pass
both the frontend and backend service.